client_windows/requirements.txt
customtkinter==5.2.1
requests==2.31.0
psutil==5.9.6
WMI==1.5.1
pywin32==306
Pillow==10.0.1
pyperclip==1.8.2